refactor(java:S1905): code smell - remove unnecessary cast to long
https://sonarcloud.io/organizations/paranoiduser/rules?open=java%3AS1905&rule_key=java%3AS1905

+++ kata/7-kyu/functions-of-integers-on-cartesian-plane/main/Funcij.java
@@ -2,11 +2,11 @@ import java.math.BigInteger;
 
 interface Funcij {
   static BigInteger sumin(int n) {
-    return BigInteger.valueOf((long) n * (n + 1) * (long) (2 * n + 1) / 6);
+    return BigInteger.valueOf((long) n * (n + 1) * (2 * n + 1) / 6);
   }
 
   static BigInteger sumax(int n) {
-    return BigInteger.valueOf((long) n * n * n - (long) n * (n - 1) * (long) (2 * n - 1) / 6);
+    return BigInteger.valueOf((long) n * n * n - (long) n * (n - 1) * (2 * n - 1) / 6);
   }
 
   static BigInteger sumsum(int n) {
